Work on type errors in reactive_codeplanner
