+++ kai/reactive_codeplanner/task_runner/dependency/api.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from kai.reactive_codeplanner.api import ValidationError
+from kai.reactive_codeplanner.task_manager.api import ValidationError
 
 
 @dataclass
