+++ kai/reactive_codeplanner/tests/test_priority_queue.py
@@ -1,7 +1,7 @@
 import unittest
 
 # Import classes from your codebase
-from kai.reactive_codeplanner.api import (
+from kai.reactive_codeplanner.task_manager.api import (
     RpcClientConfig,
     Task,
     TaskResult,
